+++ src/main/java/de/sldk/mc/config/MetricConfig.java
@@ -7,10 +7,12 @@ import java.util.function.Function;
 
 public class MetricConfig extends PluginConfig<Boolean> {
 
+    private static final String CONFIG_PATH_PREFIX = "enable_metrics";
+
     private Function<Plugin, Metric> metricInitializer;
 
     protected MetricConfig(String key, Boolean defaultValue, Function<Plugin, Metric> metricInitializer) {
-        super(key, defaultValue);
+        super(CONFIG_PATH_PREFIX + "." + key, defaultValue);
         this.metricInitializer = metricInitializer;
     }