+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased - 2019-07-15
+### Changed
+- [Breaking] Metric `mc_players_total` exports unique players. Use `mc_players_online_total` for online player count.
 
+### Added
+- Metrics can be enabled individually via config `enable_metrics`
 
 ## [v1.3.0] - 2019-05-28
 ### Added